+++ test/cljgrapht/io_test.clj
@@ -1,0 +1,48 @@
+(ns cljgrapht.io-test
+  (:require [clojure.java.io :as jio]
+            [clojure.test :refer [deftest testing is]]
+            [cljgrapht.core :as g]
+            [cljgrapht.io :as gio]))
+
+(defn- undirected-edge-set [gr]
+  (set (map set (g/edges gr))))
+
+(deftest dot-export-and-import
+  (let [gr (g/graph [[:a :b] [:b :c]])
+        s (gio/dot gr)
+        imported (gio/read-dot s)]
+    (testing "exports DOT with labels for Clojure vertex values"
+      (is (string? s))
+      (is (re-find #"graph G" s))
+      (is (re-find #"label=\":a\"" s)))
+    (testing "imports DOT into string vertices"
+      (is (= 3 (count (g/vertices imported))))
+      (is (= 2 (count (g/edges imported))))
+      (is (= #{#{"_a" "_b"} #{"_b" "_c"}}
+             (undirected-edge-set imported))))))
+
+(deftest dot-write-and-import-file
+  (let [gr (g/digraph [[:a :b]])
+        f (java.io.File/createTempFile "cljgrapht" ".dot")]
+    (try
+      (gio/write-dot! gr (.getPath f))
+      (let [imported (gio/read-dot (.getPath f))]
+        (is (= #{"_a" "_b"} (g/vertices imported)))
+        (is (= #{["_a" "_b"]} (set (g/edges imported)))))
+      (finally
+        (.delete f)))))
+
+(deftest graphml-export-and-write
+  (let [gr (g/graph [[:a :b]])
+        s (gio/graphml gr)
+        f (java.io.File/createTempFile "cljgrapht" ".graphml")]
+    (try
+      (testing "exports GraphML"
+        (is (string? s))
+        (is (re-find #"<graphml" s))
+        (is (re-find #"<node" s)))
+      (testing "writes GraphML"
+        (gio/write-graphml! gr (.getPath f))
+        (is (= s (slurp (jio/file f)))))
+      (finally
+        (.delete f)))))